requirements.txt
streamlit==1.30.0
streamlit_option_menu==0.3.6
opencv-python==4.9.0.80
ultralytics==8.0.196
pillow==10.4.0
numpy==1.26.4
pandas==1.5.3
pathlib==1.0.1
playsound==1.2.2
altair==4.2.2
matplotlib==3.8.4
seaborn==0.12.2